file split; minor doc addition
