--- src/specializations.rs
@@ -1,6 +1,0 @@
-mod single_list;
-mod tuple_of_lists;
-mod vec_of_lists;
-
-// re-export
-pub use self::single_list::Repeated;
